requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
psycopg2-binary==2.9.11
python-decouple==3.8
django-filter==25.2
Pillow==12.1.0
stripe==14.2.0
openai==2.15.0
celery==5.6.2
redis==7.1.0
python-dotenv==1.2.1
django-celery-beat==2.8.1
requests==2.32.5
pytesseract==0.3.13
bcrypt==4.2.2
pyotp==2.9.0
gunicorn==23.0.0
